+++ tests/Unit/Binding/GetDataTypeTest.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Tests\Unit\Binding;
+
+use PDO;
+use Stratedge\Wye\Binding;
+use Stratedge\Wye\Wye;
+
+class GetDataTypeTest extends \Tests\TestCase
+{
+    public function testRetrievesDataType()
+    {
+        $binding = new Binding(new Wye, 'truck', 'Engine', PDO::PARAM_STR);
+
+        $this->assertSame(PDO::PARAM_STR, $binding->getDataType());
+    }
+}