+++ tests/Unit/Binding/IsBooleanTest.php
@@ -1,0 +1,38 @@
+<?php
+
+namespace Tests\Unit\Binding;
+
+use PDO;
+use Stratedge\Wye\Binding;
+use Stratedge\Wye\Wye;
+
+class IsBooleanTest extends \Tests\TestCase
+{
+    public function testReturnsTrue()
+    {
+        $binding = new Binding(new Wye, 'truck', 'Engine', PDO::PARAM_BOOL);
+
+        $this->assertTrue($binding->isBoolean());
+    }
+
+    public function testReturnsFalse()
+    {
+        $binding = new Binding(new Wye, 'truck', 'Engine', PDO::PARAM_INT);
+
+        $this->assertFalse($binding->isBoolean());
+    }
+
+    public function testPartOfInputOutputReturnsTrue()
+    {
+        $binding = new Binding(new Wye, 'truck', 'Engine', PDO::PARAM_BOOL | PDO::PARAM_INPUT_OUTPUT);
+
+        $this->assertTrue($binding->isBoolean());
+    }
+
+    public function testNotPartOfInputOutputReturnsFalse()
+    {
+        $binding = new Binding(new Wye, 'truck', 'Engine', PDO::PARAM_INT | PDO::PARAM_INPUT_OUTPUT);
+
+        $this->assertFalse($binding->isBoolean());
+    }
+}