+++ tests/Unit/Binding/IsLargeObjectTest.php
@@ -1,0 +1,40 @@
+<?php
+
+namespace Tests\Unit\Binding;
+
+use PDO;
+use Stratedge\Wye\Binding;
+use Stratedge\Wye\Wye;
+
+class IsLargeObjectTest extends \Tests\TestCase
+{
+    public function testReturnsTrue()
+    {
+        $binding = new Binding(new Wye, 'truck', 'Engine', PDO::PARAM_LOB);
+
+        $this->assertTrue($binding->isLargeObject());
+    }
+
+    public function testReturnsFalse()
+    {
+        $binding = new Binding(new Wye, 'truck', 'Engine', PDO::PARAM_BOOL);
+
+        $this->assertFalse($binding->isLargeObject());
+    }
+
+    public function testPartOfInputOutputReturnsTrue()
+    {
+        $binding = new Binding(new Wye, 'truck', 'Engine', PDO::PARAM_LOB | PDO::PARAM_INPUT_OUTPUT);
+
+        $this->assertTrue($binding->isLargeObject());
+    }
+
+    public function testNotPartOfInputOutputReturnsFalse()
+    {
+        $binding = new Binding(new Wye, 'truck', 'Engine', PDO::PARAM_BOOL | PDO::PARAM_INPUT_OUTPUT);
+
+        $this->assertFalse($binding->isLargeObject());
+    }
+}
+
+